pyproject.toml
[tool.poetry]
name = "duckdb-demo"
version = "0.1.0"
description = "Demo of duckdb"
authors = ["Dat Nguyen <dat@infinitelambda>"]
license = "MIT"
readme = "README.md"
packages = [{include = "duckdb_demo"}]

[tool.poetry.dependencies]
python = "^3.9,!=3.9.7"
dbt-core = "^1.4.5"
dbt-duckdb = "^1.4.1"
snowflake-connector-python = "^3.0.1"
pandas = "^1.5.3"
streamlit = "^1.20.0"
plotly = "^5.13.1"
duckdb = "^0.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
